fix results types

+++ Interface web/node_scripts/database/migrations/20210306201339-create-copy.js
@@ -15,7 +15,7 @@ module.exports = {
           type: DataTypes.STRING,
       },
       result:{
-        type: DataTypes.ARRAY(DataTypes.INTEGER),
+        type: DataTypes.ARRAY(DataTypes.FLOAT),
       },
       answers:{
         type: DataTypes.STRING(2048),
